+++ src/components/Message/__tests__/MessageOptions.test.js
@@ -9,7 +9,6 @@ import { MessageSimple } from '../MessageSimple';
 import { ACTIONS_NOT_WORKING_IN_THREAD, MESSAGE_ACTIONS } from '../utils';
 
 import { Attachment } from '../../Attachment';
-import { MessageActions as MessageActionsMock } from '../../MessageActions';
 
 import { ChannelActionProvider } from '../../../context/ChannelActionContext';
 import { ChannelStateProvider } from '../../../context/ChannelStateContext';
@@ -23,9 +22,7 @@ import {
   getTestClientWithUser,
 } from '../../../mock-builders';
 
-jest.mock('../../MessageActions', () => ({
-  MessageActions: jest.fn(() => <div />),
-}));
+const MESSAGE_ACTIONS_TEST_ID = 'message-actions';
 
 const minimumCapabilitiesToRenderMessageActions = { 'delete-any-message': true };
 const alice = generateUser({ name: 'alice' });
@@ -185,43 +182,45 @@ describe('<MessageOptions />', () => {
   });
 
   it('should render message actions', async () => {
-    await renderMessageOptions({
+    const { queryByTestId } = await renderMessageOptions({
       channelStateOpts: { channelCapabilities: minimumCapabilitiesToRenderMessageActions },
     });
-    // eslint-disable-next-line jest/prefer-called-with
-    expect(MessageActionsMock).toHaveBeenCalled();
+
+    expect(queryByTestId(MESSAGE_ACTIONS_TEST_ID)).toBeInTheDocument();
   });
 
   it('should not show message actions button if actions are disabled', async () => {
-    await renderMessageOptions({
+    const { queryByTestId } = await renderMessageOptions({
       channelStateOpts: { channelCapabilities: minimumCapabilitiesToRenderMessageActions },
       customMessageProps: { messageActions: [] },
     });
-    expect(MessageActionsMock).not.toHaveBeenCalled();
+
+    expect(queryByTestId(MESSAGE_ACTIONS_TEST_ID)).not.toBeInTheDocument();
   });
 
   it('should not show actions box for message in thread if only non-thread actions are available', async () => {
-    await renderMessageOptions({
+    const { queryByTestId } = await renderMessageOptions({
       channelStateOpts: { channelCapabilities: minimumCapabilitiesToRenderMessageActions },
       customMessageProps: { messageActions: ACTIONS_NOT_WORKING_IN_THREAD, threadList: true },
     });
-    expect(MessageActionsMock).not.toHaveBeenCalled();
+
+    expect(queryByTestId(MESSAGE_ACTIONS_TEST_ID)).not.toBeInTheDocument();
   });
 
   it('should show actions box for message in thread if not only non-thread actions are available', async () => {
-    await renderMessageOptions({
+    const { queryByTestId } = await renderMessageOptions({
       channelStateOpts: { channelCapabilities: minimumCapabilitiesToRenderMessageActions },
       customMessageProps: {
         messageActions: [...ACTIONS_NOT_WORKING_IN_THREAD, MESSAGE_ACTIONS.delete],
         threadList: true,
       },
     });
-    // eslint-disable-next-line jest/prefer-called-with
-    expect(MessageActionsMock).toHaveBeenCalled();
+
+    expect(queryByTestId(MESSAGE_ACTIONS_TEST_ID)).toBeInTheDocument();
   });
 
   it('should show actions box for a message in thread if custom actions provided are non-thread', async () => {
-    await renderMessageOptions({
+    const { queryByTestId } = await renderMessageOptions({
       channelStateOpts: { channelCapabilities: minimumCapabilitiesToRenderMessageActions },
       customMessageProps: {
         customMessageActions: ACTIONS_NOT_WORKING_IN_THREAD,
@@ -229,78 +228,76 @@ describe('<MessageOptions />', () => {
         threadList: true,
       },
     });
-    // eslint-disable-next-line jest/prefer-called-with
-    expect(MessageActionsMock).toHaveBeenCalled();
+    expect(queryByTestId(MESSAGE_ACTIONS_TEST_ID)).toBeInTheDocument();
   });
 
   it('should not show actions box for message outside thread with single action "react"', async () => {
-    await renderMessageOptions({
+    const { queryByTestId } = await renderMessageOptions({
       channelStateOpts: { channelCapabilities: minimumCapabilitiesToRenderMessageActions },
       customMessageProps: {
         messageActions: [MESSAGE_ACTIONS.react],
       },
     });
-    // eslint-disable-next-line jest/prefer-called-with
-    expect(MessageActionsMock).not.toHaveBeenCalled();
+    expect(queryByTestId(MESSAGE_ACTIONS_TEST_ID)).not.toBeInTheDocument();
   });
 
   it('should show actions box for message outside thread with single action "react" if custom actions available', async () => {
-    await renderMessageOptions({
+    const { queryByTestId } = await renderMessageOptions({
       channelStateOpts: { channelCapabilities: minimumCapabilitiesToRenderMessageActions },
       customMessageProps: {
         customMessageActions: [MESSAGE_ACTIONS.react],
         messageActions: [MESSAGE_ACTIONS.react],
       },
     });
-    // eslint-disable-next-line jest/prefer-called-with
-    expect(MessageActionsMock).toHaveBeenCalled();
+
+    expect(queryByTestId(MESSAGE_ACTIONS_TEST_ID)).toBeInTheDocument();
   });
 
   it('should not show actions box for message outside thread with single action "reply"', async () => {
-    await renderMessageOptions({
+    const { queryByTestId } = await renderMessageOptions({
       channelStateOpts: { channelCapabilities: minimumCapabilitiesToRenderMessageActions },
       customMessageProps: {
         messageActions: [MESSAGE_ACTIONS.reply],
       },
     });
-    // eslint-disable-next-line jest/prefer-called-with
-    expect(MessageActionsMock).not.toHaveBeenCalled();
+
+    expect(queryByTestId(MESSAGE_ACTIONS_TEST_ID)).not.toBeInTheDocument();
   });
 
   it('should show actions box for message outside thread with single action "reply" if custom actions available', async () => {
-    await renderMessageOptions({
+    const { queryByTestId } = await renderMessageOptions({
       channelStateOpts: { channelCapabilities: minimumCapabilitiesToRenderMessageActions },
       customMessageProps: {
         customMessageActions: [MESSAGE_ACTIONS.reply],
         messageActions: [MESSAGE_ACTIONS.reply],
       },
     });
-    // eslint-disable-next-line jest/prefer-called-with
-    expect(MessageActionsMock).toHaveBeenCalled();
+
+    expect(queryByTestId(MESSAGE_ACTIONS_TEST_ID)).toBeInTheDocument();
   });
 
   it('should not show actions box for message outside thread with two actions "react" & "reply"', async () => {
     const actions = [MESSAGE_ACTIONS.react, MESSAGE_ACTIONS.reply];
-    await renderMessageOptions({
+    const { queryByTestId } = await renderMessageOptions({
       channelStateOpts: { channelCapabilities: minimumCapabilitiesToRenderMessageActions },
       customMessageProps: {
         messageActions: actions,
       },
     });
-    // eslint-disable-next-line jest/prefer-called-with
-    expect(MessageActionsMock).not.toHaveBeenCalled();
+
+    expect(queryByTestId(MESSAGE_ACTIONS_TEST_ID)).not.toBeInTheDocument();
   });
 
   it('should show actions box for message outside thread with single actions "react" & "reply" if custom actions available', async () => {
     const actions = [MESSAGE_ACTIONS.react, MESSAGE_ACTIONS.reply];
-    await renderMessageOptions({
+    const { queryByTestId } = await renderMessageOptions({
       channelStateOpts: { channelCapabilities: minimumCapabilitiesToRenderMessageActions },
       customMessageProps: {
         customMessageActions: actions,
         messageActions: actions,
       },
     });
-    // eslint-disable-next-line jest/prefer-called-with
-    expect(MessageActionsMock).toHaveBeenCalled();
+
+    expect(queryByTestId(MESSAGE_ACTIONS_TEST_ID)).toBeInTheDocument();
   });
 });